+++ EVTCAnalytics/Events/GadgetCaptureEvent.cs
@@ -31,12 +31,12 @@ namespace GW2Scratch.EVTCAnalytics.Events
 	/// <remarks>
 	/// Introduced in EVTC20260610.
 	/// </remarks>
-	public class GadgetCaptureSplitPercentEvent(long time, Agent agent, int percentage, byte cappingFrom, byte cappingBy) : GadgetCaptureEvent(time, agent)
+	public class GadgetCaptureSplitPercentEvent(long time, Agent agent, float percentage, byte cappingFrom, byte cappingBy) : GadgetCaptureEvent(time, agent)
 	{
 		/// <summary>
 		/// Capture percentage.
 		/// </summary>
-		public int Percentage { get; } = percentage;
+		public float Percentage { get; } = percentage;
 		/// <summary>
 		/// Being captured from [color id number].
 		/// </summary>
@@ -56,5 +56,27 @@ namespace GW2Scratch.EVTCAnalytics.Events
 	public class GadgetCaptureOutlineHideEvent(long time, Agent agent) : GadgetCaptureEvent(time, agent)
 	{
 
+	}
+
+	/// <summary>
+	/// Outline point event. There will be one event for each point of the outline.
+	/// </summary>
+	/// <remarks>
+	/// Introduced in EVTC20260612.
+	/// </remarks>
+	public class GadgetCaptureOutlinePointEvent(long time, Agent agent, int pointIndex, float[] position, uint pointCount) : GadgetCaptureEvent(time, agent)
+	{
+		/// <summary>
+		/// The index of the point.
+		/// </summary>
+		public int PointIndex { get; } = pointIndex;
+		/// <summary>
+		/// Position of the point of the outline.
+		/// </summary>
+		public float[] Position { get; } = position;
+		/// <summary>
+		/// The number of points of the capture outline.
+		/// </summary>
+		public uint PointCount { get; } = pointCount;
 	}
 }